api/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
openai
azure-data-tables

pinecone
langchain-pinecone
langchain-openai
langchain-core
langgraph
python-dotenv